\name{FOCUS_2006_B}
\Rdversion{1.1}
\alias{FOCUS_2006_B}
\docType{data}
\title{
Dataset B from the FOCUS Kinetics report from 2006
}
\description{
Data generated using a model and assuming some variability.
}
\usage{data(FOCUS_2006_B)}
\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{t}}{a numeric vector containing time points}
    \item{\code{parent}}{a numeric vector containing parent concentrations in percent of applied radioactivity}
  }
}
\source{
  FOCUS (2006) \dQuote{Guidance Document on Estimating Persistence and
  Degradation Kinetics from Environmental Fate Studies on Pesticides in EU
  Registration} Report of the FOCUS Work Group on Degradation Kinetics,
  EC Document Reference Sanco/10058/2005 version 2.0, 434 pp,
  \url{http://focus.jrc.ec.europa.eu/dk} 
}
\examples{
data(FOCUS_2006_B)
}
\keyword{datasets}
